fixed a bug where locale folders were ignored if both clone and download folder were off improved unit test cleanup for testing the add function

+++ tests/test_actions/test_add.py
@@ -26,9 +26,16 @@ class TestAdd(unittest.TestCase):
         for fn in self.added_files:
             self.rm_action.rm_action(fn, force=True)
 
-        for d in self.added_directories:
-            self.rm_action.rm_action(d, force=True)
-            delete_directory(d)
+        not_empty = True
+        while not_empty:
+            not_empty = False
+            for d in self.added_directories:
+                if os.path.exists(d) and os.path.isdir(d):
+                    if len(os.listdir(d)):
+                        not_empty = True
+                    else:
+                        self.rm_action.rm_action(d, force=True)
+                        delete_directory(d)
 
         self.clean_action.clean_action(False, False, None)
         self.rm_action.close()
